I read more into clone() than it does; remove it from unix chapter. Mention that OSes optimize fork for the fork/exec pattern.
It would be nice to have a paragraph in "real world" that gives some
more detailed examples where the Unix interface isn't perfect, but
didn't see an easy to bring that up.  Perhaps it should be in
subsequent chapters when we dive in more detail.

In the sched.t chapter mention that today's OSes support several
threads within a single process, and that has many challenges. Bring
clone up in that context.